Bot modifieded

+++ paper-io/Player.cs
@@ -62,10 +62,22 @@ namespace paper_io
                     return true;
             return false;
         }
-        int m; // территоря текущего икгрока
-        int k; // стена
-        int c; // свободная территория
-        int q; // вражеская территория
+        /// <summary>
+        /// Территория текущего игрока
+        /// </summary>
+        int m;
+        /// <summary>
+        /// Стена поля
+        /// </summary>
+        int k;
+        /// <summary>
+        /// свободная территория
+        /// </summary>
+        int c;
+        /// <summary>
+        /// вражеская территория
+        /// </summary>
+        int q; 
 
         public void Step(Game[,] room)
         {
@@ -80,24 +92,25 @@ namespace paper_io
         public void Bot()
         {
             ///Если со всех сторон находится территория текущего игрока, то направление движения не менять. 
-            if (location.X + 1 == m && location.X - 1 == m && location.Y + 1 == m && location.Y - 1 == m)
+            if ((int)location.X + 1 == m && (int)location.X - 1 == m && (int)location.Y + 1 == m && (int)location.Y - 1 == m)
             {
 
             }
             ///Если впереди текущего игрока находится стена и слева нет стены, то повернуть на лево.
-            else if (location.X + 1 == k && location.Y - 1 != k)
+            else if ((int)location.X + 1 == k && (int)location.Y - 1 != k)
             {
                 ToLeft();
             }
             ///Если впереди игрока находится стена и справа нет стены, то повернуть направо.
-            else if (location.X + 1 == k  && location.Y + 1 !=  k )
+            else if ((int)location.X + 1 == k  && (int)location.Y + 1 !=  k )
             {
                 ToRight();
             }
-            ///Если впереди и слева и справа нет территории текущего игрока, то повернуть направо если там нет стены или повернуть налево если справа есть стена.
-            else if (location.X + 1 == m && location.Y + 1 == m && location.Y - 1 == m)
+            ///Если впереди и слева и справа нет территории текущего игрока, 
+            ///то повернуть направо если там нет стены или повернуть налево если справа есть стена.
+            else if ((int)location.X + 1 == m && (int)location.Y + 1 == m && (int)location.Y - 1 == m)
             {
-                if (location.Y - 1 != k )
+                if ((int)location.Y - 1 != k )
                 {
                     ToRight();
                 }
@@ -106,10 +119,11 @@ namespace paper_io
                     ToLeft();
                 }
             }
-            ///Если впереди нет территории текущего игрока, а слево или справа есть территория текущего игрока, то повернуть в сторону территории текущего игрока.
-            else if (location.X + 1 != m )
+            ///Если впереди нет территории текущего игрока, а слево или справа есть территория текущего игрока, 
+            ///то повернуть в сторону территории текущего игрока.
+            else if ((int)location.X + 1 != m )
             {
-                if (location.Y + 1 == m)
+                if ((int)location.Y + 1 == m)
                 {
                     ToLeft();
                 }
